Move confirm missing email fragment to fragment package
